datatype

+++ codeforces/800/coins.cpp
@@ -4,7 +4,7 @@ int main(){
     int t;
     cin>>t;
     while(t--){
-        int n,k;
+        long long  n,k;
         cin>>n>>k;
         if(n%2==0 || (n-k)%2==0){
             cout<<"yes"<<endl;
